Cleaned Data
Hey all!

This is the cleaned data (I did a little in excel because it was faster
and I was working under a time constraint), the code I used to do some
of the cleaning and my analyses, and my project (it’s quite boring and
shitty, please excuse that!).