my portion: final updates until I proof read and do a final check on Sunday 3/19/17
I finished all of my parts of the assignments.  I haven’t proof read
anything  yet.  I want to give myself a couple of days to sort of
forget about the project and look at it on Sunday with a fresh set of
eyes to make my final edits.  Going forward, please add all updates to
this r markdown file.